Use long for memory limit

+++ ACS.Shared/Configuration/HttpLoggingConfiguration.cs
@@ -20,7 +20,7 @@ namespace ACS.Shared.Configuration
         /// <summary>
         /// Size of in-memory queue (where persistence is disabled)
         /// </summary>
-        public int? InMemoryQueueLimitBytes { get; set; }
+        public long? InMemoryQueueLimitBytes { get; set; }
 
         public PersistenceConfiguration? Persistence { get; set; }
